+++ lib/wasm-interface/src/interface.rs
@@ -4,6 +4,8 @@ use std::collections::HashMap;
 
 #[derive(Debug, Clone, PartialEq, Eq, Default)]
 pub struct Interface {
+    /// The name the interface gave itself
+    pub name: Option<String>,
     /// Things that the module can import
     pub imports: HashMap<(String, String), Import>,
     /// Things that the module must export
@@ -129,13 +131,17 @@ mod test {
 
     #[test]
     fn merging_works() {
-        let interface1_src = r#"(assert_import (func "env" "plus_one" (param i32) (result i32)))"#;
-        let interface2_src = r#"(assert_import (func "env" "plus_one" (param i64) (result i64)))"#;
-        let interface3_src = r#"(assert_import (func "env" "times_two" (param i64) (result i64)))"#;
+        let interface1_src =
+            r#"(interface (func (import "env" "plus_one") (param i32) (result i32)))"#;
+        let interface2_src =
+            r#"(interface (func (import "env" "plus_one") (param i64) (result i64)))"#;
+        let interface3_src =
+            r#"(interface (func (import "env" "times_two") (param i64) (result i64)))"#;
         let interface4_src =
-            r#"(assert_import (func "env" "times_two" (param i64 i64) (result i64)))"#;
-        let interface5_src = r#"(assert_export (func "empty_bank_account" (param) (result)))"#;
-        let interface6_src = r#"(assert_export (func "empty_bank_account" (param) (result i64)))"#;
+            r#"(interface (func (import "env" "times_two") (param i64 i64) (result i64)))"#;
+        let interface5_src = r#"(interface (func (export "empty_bank_account") (param) (result)))"#;
+        let interface6_src =
+            r#"(interface (func (export "empty_bank_account") (param) (result i64)))"#;
 
         let interface1 = parser::parse_interface(interface1_src).unwrap();
         let interface2 = parser::parse_interface(interface2_src).unwrap();