+++ src/constants.rs
@@ -1,0 +1,4 @@
+pub const DEFAULT_RUNTIME: &str = "wasmer";
+
+pub const RFC3339_FORMAT_STRING: &'static str = "%Y-%m-%dT%H:%M:%S-%f";
+pub const RFC3339_FORMAT_STRING_WITH_TIMEZONE: &'static str = "%Y-%m-%dT%H:%M:%S.%f+%Z";